requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
minio==7.2.7
celery==5.4.0
redis==5.0.8
kafka-python==2.0.2
pika==1.3.2
websockets==13.1
pytest==8.3.3
httpx==0.27.0
black==24.8.0
pypdf==4.3.1
python-docx==1.1.2  # For docx→PDF
img2pdf==1.17.0     # For image→PDF
qrcode[pil]==7.4.2
Pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
opencv-python==4.10.0

opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.5
xlsxwriter==3.2.0
pydub==0.25.1


matplotlib==3.9.2
seaborn==0.13.2
scipy==1.14.1
plotly==5.24.1
kaleido==0.3.0  # For static chart exports
celery[redis]==5.4.0
nltk==3.8.1
kafka-python==2.0.2
pika==1.3.2
aio-pika==9.4.1
